requirements.txt
django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2